Move MySQLStatementIDGenerator from protocol to frontend
